factor out waiver table
